WAB::Dev started
Set up Impl module and testing

Add the Impl module along with updated to teh WAB::Data class. The
start of tests is included. Uses the new Oj :wab mode which is in
release 3.3.0.

Validate and fix Data class

trigger travis

Added rake as a dependency in gemspec

Added rake as a dependency in gemspec

another travei attempt

downgrading to 2.4.0 for osx on travis

downgrading to 2.4.0 for osx on travis

finished off data tests